[#34] Firebase storage접근 의존성 주입
